+++ docs/changelog.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - `force_large` argument in the `pull` method, which can be used to handle large archive downloads
 - `add` method
 
+### Changed
+- `getseq` method returns the sequence string instead of printing it to the screen
+
 ### Deprecated
 - `get_remote_data_str` method. Use `listr` instead